Common functions moved to D3Files component
