+++ widgets/D3FilesWidget.php
@@ -2,7 +2,8 @@
 
 namespace d3yii2\d3files\widgets;
 
-use d3yii2\d3files\D3Files;
+use d3yii2\d3files\components\D3Files;
+use d3yii2\d3files\D3Files as D3FilesModule;
 use d3yii2\d3files\models\D3files as ModelD3Files;
 use Exception;
 use Yii;
@@ -79,7 +80,7 @@ class D3FilesWidget extends Widget
      */
     public function init(): void
     {
-        D3Files::registerTranslations();
+        D3FilesModule::registerTranslations();
 
         if (property_exists($this->model, 'd3filesControllerRoute')) {
             $this->controllerRoute = $this->model->d3filesControllerRoute;
@@ -101,10 +102,6 @@ class D3FilesWidget extends Widget
         // Load the file list if has not been set in constructor
         if (!$this->fileList) {
             $this->fileList = ModelD3Files::fileListForWidget($this->model_name, $this->model_id);
-        }
-
-        if (!$this->readOnly) {
-            $this->registerJsTranslations();
         }
     }
 
@@ -169,50 +166,6 @@ class D3FilesWidget extends Widget
         return $this->fileList;
     }
 
-    /**
-     * Get file extension
-     * @param array $file
-     * @return string
-     */
-    public static function getFileExtension(array $file): string
-    {
-        return strtolower(pathinfo($file['file_name'], PATHINFO_EXTENSION));
-    }
-
-    /**
-     * Return the first file of the fileList having extension
-     * @param array $files
-     * @param string $extension
-     * @return array|null
-     */
-    public static function getFirstFileHavingExt(array $files, string $extension): array
-    {
-        foreach ($files as $file) {
-            $fileExtension = strtolower(pathinfo($file['file_name'], PATHINFO_EXTENSION));
-            if ($extension === $fileExtension) {
-                return $file;
-            }
-        }
-        return [];
-    }
-
-    /**
-     * Filter the files by extension
-     * @param array $files
-     * @param string $ext
-     * @return array
-     */
-    public static function getFilesListByExt(array $files, string $ext): array
-    {
-        $list = [];
-        foreach ($files as $file) {
-            $fileExt = self::getFileExtension($file);
-            if ($ext === $fileExt) {
-                $list[$file['id']] = $file;
-            }
-        }
-        return $list;
-    }
 
     public function registerJsTranslations()
     {
